Initial implementation of a Queue::Policy
