+++ lib/postqueue.rb
@@ -2,6 +2,7 @@ require_relative "postqueue/logger"
 require_relative "postqueue/item"
 require_relative "postqueue/version"
 require_relative "postqueue/queue"
+require_relative "postqueue/policy"
 require_relative "postqueue/default_queue"
 require_relative "postqueue/availability"
 
@@ -14,6 +15,23 @@ module Postqueue
       raise ArgumentError, "Invalid table_name parameter" unless table_name
 
       ::Postqueue::Queue.new(table_name: table_name)
+    end
+
+    def queue_for_table(table_name)
+      @queues_for_table ||= {}
+      @queues_for_table[table_name] ||= new(table_name: table_name)
+    end
+
+    def run!(table_name: DEFAULT_TABLE_NAME)
+      queue_for_table(table_name).run!
+    end
+
+    def migrate!(table_name: DEFAULT_TABLE_NAME)
+      queue_for_table(table_name).item_class.migrate!
+    end
+
+    def unmigrate!(table_name: DEFAULT_TABLE_NAME)
+      queue_for_table(table_name).item_class.unmigrate!
     end
   end
 end
