+++ lib/postqueue/queue/exception_handling.rb
@@ -14,18 +14,18 @@ module Postqueue
   end
 
   module ExceptionHandling
-    private
-
-    def log_exception(exception, op, entity_ids)
-      Postqueue.logger.warn "processing '#{op}' for id(s) #{entity_ids.inspect}: caught #{exception}"
-    end
-
     def on_exception(&block)
       raise ArgumentError, "on_exception expects a block argument" unless block
       raise ArgumentError, "on_exception expects a block accepting 3 or more arguments" if block.arity > -3 && block.arity != 3
 
       @on_exception = block
       self
+    end
+
+    private
+
+    def log_exception(exception, op, entity_ids)
+      Postqueue.logger.warn "processing '#{op}' for id(s) #{entity_ids.inspect}: caught #{exception}"
     end
   end
